Change sig for daemonize to unsigned

+++ gx_net.h
@@ -88,12 +88,12 @@ gx_tcp_listen_err:
     _raise(-1);
 }
 
-static inline int gx_net_daemonize(void) {
+static inline unsigned int gx_net_daemonize(void) {
     /* TODO: chroot option? */
     /* TODO: it should be possible to make this much lighter with clone. */
     pid_t pid, sid;
     _ (pid = fork()) _raise_error(-1);
-    if(pid != 0) return pid; //exit(EXIT_SUCCESS);
+    if(pid != 0) return (unsigned int)pid; //exit(EXIT_SUCCESS);
     /* Child */
     umask(0);
     _ (sid = setsid()) _warning();
